update to List<byte> for better performance and less space taken

+++ lib/server/data_model/budget_list.dart
@@ -19,7 +19,7 @@ class BudgetList {
 
   final category = IsarLink<Category>();
 
-  late List<int> imagesBytes;
+  late List<byte> imagesBytes;
 
   @Index(
     unique: false,
